Remove attestations file from @celo/base
