+++ packages/phone-number-privacy/common/src/test/utils.ts
@@ -1,4 +1,3 @@
-import { AttestationsStatus } from '@celo/base'
 import { privateKeyToAddress } from '@celo/utils/lib/address'
 import { serializeSignature, Signature, signMessage } from '@celo/utils/lib/signatureUtils'
 import BigNumber from 'bignumber.js'
@@ -13,6 +12,13 @@ import {
 } from '../interfaces'
 import { signWithRawKey } from '../utils/authentication'
 import { genSessionID } from '../utils/logger'
+
+export interface AttestationsStatus {
+  isVerified: boolean
+  numAttestationsRemaining: number
+  total: number
+  completed: number
+}
 
 export function createMockAttestation(getVerifiedStatus: jest.Mock<AttestationsStatus, []>) {
   return {
