+++ packages/sdk/base/src/index.ts
@@ -1,7 +1,6 @@
 export * from './account'
 export * from './address'
 export * from './async'
-export * from './attestations'
 export * from './collections'
 export * from './contacts'
 export * from './currencies'
